requirements.txt
streamlit
streamlit-option-menu
streamlit-mic-recorder
torch -f https://download.pytorch.org/whl/cu118/torch_stable.html
torchvision -f https://download.pytorch.org/whl/cu118/torch_stable.html
torchaudio -f https://download.pytorch.org/whl/cu118/torch_stable.html
ctransformers
transformers
langchain
langchain-community
InstructorEmbedding
sentence-transformers==2.2.2
faiss-cpu
python-docx
pypdf
underthesea
pyyaml